Reorder pick and roll and fix skeleton
