+++ app/dashboard/pick/page.tsx
@@ -13,6 +13,22 @@ export default function Page() {
   const [filterValue, setFilterValue] = useState<string>('all')
   const [isLoading, setIsLoading] = useState(true)
 
+  // Random food truck
+  const [randomFoodTrucks, setRandomFoodTrucks] = useState<FoodTruck[]>([])
+
+  // Function to select a random food truck
+  const selectRandomFoodTrucks = () => {
+    const numberOfRandomTrucks = 3
+    const randomTrucks = []
+
+    for (let i = 0; i < numberOfRandomTrucks; i++) {
+      const randomIndex = Math.floor(Math.random() * filteredTrucks.length)
+      randomTrucks.push(filteredTrucks[randomIndex])
+    }
+
+    setRandomFoodTrucks(randomTrucks)
+  }
+
   useEffect(() => {
     // setIsClient(true)
     axios
@@ -51,55 +67,38 @@ export default function Page() {
   return (
     <div className="w-full h-full text-center p-10">
       <div className="flex border-fresh-textColor justify-center"></div>
+      {/* <p className="italic text-gray-400">A local food truck service</p> */}
+      <div className="mt-20">
+        <button
+          onClick={selectRandomFoodTrucks}
+          className="bg-fresh-textColor hover:bg-fresh-subTextColor text-[#ffffff] text-xl font-bold py-4 px-6 rounded"
+        >
+          Roll Three For Me
+        </button>
+        <p className="text-gray-400 mt-3">
+          Hard to decide what to eat, roll for me.
+        </p>
+      </div>
       <div className="w-full h-full text-center p-10">
-        <div className="mt-5 mb-10">
-          <h2 className="font-bold text-xl mb-5 text-textColor">
-            Pick lunch idea by country
-          </h2>
-          <select
-            id="filterSelect"
-            className="block w-full rounded py-2 px-3"
-            value={filterValue}
-            onChange={handleFilterChange}
-          >
-            <option value="all">All</option>
-            <option value="chinese">Chinese</option>
-            <option value="mexican">Mexican</option>
-            <option value="italian">Italian</option>
-          </select>
-        </div>
         <div className="food-truck-list grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-          {isLoading
-            ? Array(9)
-                .fill(undefined)
-                .map((index) => (
-                  <div
-                    className="food-truck-card p-6 rounded shadow-lg"
-                    key={index}
-                  >
-                    <Skeleton />
-                  </div>
-                ))
-            : filteredTrucks.map((truck) => (
-                <div
-                  className="food-truck-card p-6 bg-fresh-cardColor border border-fresh-subTextColor rounded shadow-md"
-                  key={truck.objectid}
-                >
-                  <h2 className="capitalize text-2xl font-semibold text-textColor mb-8">
-                    {truck.applicant}
-                  </h2>
-                  <div className="text-left text-sm mt-2 text-gray">
-                    <p className="mb-1">Facility Type: {truck.facilitytype}</p>
-                    <p className="mb-1">
-                      Location: {truck.locationdescription}
-                    </p>
-                    <p className="mb-1">Address: {truck.address}</p>
-                    <p className="mb-1">Permit: {truck.permit}</p>
-                    <p className="mb-1">Status: {truck.status}</p>
-                    <p>Food Items: {truck.fooditems}</p>
-                  </div>
-                </div>
-              ))}
+          {randomFoodTrucks.map((truck) => (
+            <div
+              className="border-t-[30px] border border-fresh-textColor food-truck-card p-6 mb-10  bg-fresh-cardColor rounded shadow-sm"
+              key={truck.objectid}
+            >
+              <h2 className="text-2xl font-semibold text-textColor mb-8">
+                {truck.applicant}
+              </h2>
+              <div className="text-left text-sm">
+                <p className="mb-1">Facility Type: {truck.facilitytype}</p>
+                <p className="mb-1">Location: {truck.locationdescription}</p>
+                <p className="mb-1">Address: {truck.address}</p>
+                <p className="mb-1">Permit: {truck.permit}</p>
+                <p className="mb-1">Status: {truck.status}</p>
+                <p>Food Items: {truck.fooditems}</p>
+              </div>
+            </div>
+          ))}
         </div>
       </div>
     </div>
