+++ app/page.tsx
@@ -15,6 +15,7 @@ export default function Page() {
         <div className="flex flex-col justify-center gap-6 rounded-lg bg-fresh-cardColor px-6 py-10 md:w-2/5 md:px-20">
           <p className={`text-xl text-gray-800 md:text-3xl md:leading-normal`}>
             <strong>Welcome to Truckee.</strong> A local food truck service.{' '}
+            <br />
             <a
               href="https://cyfyifanchen.com"
               className="text-xl text-fresh-textColor hover:underline"
